Added edits from Dave:
o	Make it say "11.4.2+" instead of 11.4.2
o	The beginning says "This topic describes … " and then there's the "Working with PKS" section that points to the online docs. I found that confusing. I think a better approach would be to change the "This topic…" to make it clear that this topic/section is for PCF – if you're using PKS, you should see the next section. And put a link to the next section on PKS (which already has a link to the online docs)